migrate genseqTXs, add TX fuzzer
